requirements.txt
beautifulsoup4==4.4.1
Django==1.8
django-bootstrap3==6.1.0
django-chosenadmin==0.2.0
django-ckeditor==4.4.8
django-cleanup==0.4.2
django-countries==3.3
django-crispy-forms==1.4.0
django-easy-pdf==0.1.0
django-endless-pagination-angular==1.1
## FIXME: could not find svn URL in dependency_links for this package:
django-files-widget===0.2.4dev-r0
django-fixture-magic==0.0.8
django-flatpages-i18n==0.4.1
django-fsm-widget==0.0.1
django-grappelli==2.6.5
django-location-field==1.5.2
django-mailchimp-v1.3==1.3.1
django-maskurl==0.1
django-modeltranslation==0.9.1
django-mptt==0.7.4
django-nocaptcha-recaptcha==0.0.18
django-orphaned==0.1
django-pagination-bootstrap==1.0.7
django-pure-pagination==0.3.0
django-randomslugfield==0.3.0
django-realtime==1.1
django-registration-redux==1.2



django-select2-forms==1.1.25



django-settings-export==1.1.0
django-sortedm2m==1.0.1
django-unused-media==0.1.3
django-wysiwyg-redactor==0.4.9
docopt==0.4.0
easy-thumbnails==2.3
ecdsa==0.13
elasticsearch==1.6.0
Fabric==1.10.2
html5lib==0.999999
keyring==8.7
linecache2==1.0.0
lxml==3.6.0
migrate==0.3.7
MySQL-python==1.2.4
paramiko==1.16.0
pdfkit==0.5.0
Pillow==2.8.2
psycopg2==2.6.1
pycrypto==2.6.1
PyMySQL==0.6.6
PyPDF2==1.25.1
python-gettext==2.1
pytz==2015.4
pywin32-ctypes==0.0.1
reportlab==3.2.0
requests==2.7.0
simplejson==3.8.0
six==1.9.0
sorl-thumbnail==12.2
South==1.0.2
towel==0.7.0
traceback2==1.4.0
Unidecode==0.4.19
unittest2==1.1.0
urllib3==1.10.4
virtualenv==13.1.2
wheel==0.24.0
xhtml2pdf==0.0.6
yagmail==0.4.132
